Module Cohttp_curl_lwt

Curl based client

module Sink : sig ... end

A sink defines where the response body may be written

module Source : sig ... end

A source defines where the request body is read from

module Context : sig ... end

A context shares the curl event handling logic for all curl requests associated to it

module Response : sig ... end

Response for the http requests

module Request : sig ... end

Http requests

val submit : Context.t -> 'a Request.t -> 'a Response.t

submit ctx request submits a request and returns the response. Once a request is submitted, it may not be submitted again.